+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased 
 
+- Add `deployToken` method to `WalletAddress` and `Wallet` to deploy an ERC20, updated `SmartContract` class to support deployment and fetching contract details
+
 ## [0.5.0] - 2024-09-11
 
 - Add Arbitrum-Mainnet support for Native transfers.